refactor: standardize store imports to use barrel file pattern
- Updated all 22 files to import from '@/store' instead of direct file paths
- Changed imports from '@/store/useUserStore' and '@/store/userProvider' to '@/store'
- Added missing 'use client' directive to useShifu.tsx
- Ensures consistency with other store imports (useShifu, ShifuProvider)

This improves maintainability by centralizing all store exports through the barrel file, making the codebase more consistent and reducing coupling to specific file locations.
