+++ src/cook-web/src/app/c/[[...id]]/Components/NavDrawer/NavDrawer.tsx
@@ -22,7 +22,7 @@ import {
 import { useDisclosure } from '@/c-common/hooks/useDisclosure';
 import MainMenuModal from './MainMenuModal';
 
-import { useUserStore } from '@/store/useUserStore';
+import { useUserStore } from '@/store';
 
 /**
  * 导航栏展示形式
