+++ src/cook-web/src/app/c/[[...id]]/page.tsx
@@ -22,7 +22,7 @@ import {
   useUiLayoutStore,
   useSystemStore,
 } from '@/c-store';
-import { useUserStore } from '@/store/useUserStore';
+import { useUserStore } from '@/store';
 import { useDisclosure } from '@/c-common/hooks/useDisclosure';
 import { useLessonTree } from './hooks/useLessonTree';
 import { updateWxcode } from '@/c-api/user';
