+++ src/cook-web/src/components/shifu-edit/ShifuEdit.tsx
@@ -14,7 +14,7 @@ import {
   ListCollapse,
 } from 'lucide-react';
 import { useShifu } from '@/store';
-import { useUserStore } from '@/store/useUserStore';
+import { useUserStore } from '@/store';
 import OutlineTree from '@/components/outline-tree';
 import '@mdxeditor/editor/style.css';
 import Header from '../header';
