+++ src/cook-web/src/store/useShifu.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import {
   Shifu,
   ShifuContextType,
